junit 5 project - pom fixed
