+++ test/junit/junit5-project/src/test/java/TestSmoke2.java
@@ -1,13 +1,18 @@
-package example.testsuite;
-
 import java.util.concurrent.TimeUnit;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class TestSmoke2 {
     @Test
     public void testAaa() {
-        TimeUnit.SECONDS.sleep(1);
+        try {
+            TimeUnit.SECONDS.sleep(1);
+        }
+        catch(Exception e) {
+            System.out.println(e);
+        }
         assertEquals(1, 1);
     }
 
@@ -18,7 +23,12 @@ public class TestSmoke2 {
 
     @Test
     public void testCcc() {
-        TimeUnit.SECONDS.sleep(2);
+        try {
+            TimeUnit.SECONDS.sleep(2);
+        }
+        catch(Exception e) {
+            System.out.println(e);
+        }
         assertEquals(1, 1);
     }
